feat: add clean up logic for pending payloads (#52)
This PR adds the required logic to clean up pending payloads after they
have been submitted for a given lifespan period (currently set to 1
day).
The clean up is done in the `BeginBlock` and is limited to 200 payloads
per block to avoid running into timeouts or deadlocks after spam attacks
through mass payload submissions.

The iteration over the registered payloads is implemented by adding a
new `collections.Index` for the block time of submission and using an
`IndexedMap` instead of a regular `Map`. This makes it possible to call
`k.pendingPayloads.Indexes.HashByTime.Walk(...)` which has a ranger
implementation passed, to only iterate through those payloads that are
expired.